requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
joblib==1.5.2
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
gunicorn==21.2.0
